+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+Changes:
+
+  * Added Sample PacBio Revio test directory
+  * Fixed Runs not displaying for PacBio Revio Processors
+
 # 2.1.0
 
 Added: